+++ lib/backend/vm_compiler.c
@@ -250,10 +250,10 @@ static void compile_symbol_literal(FuncChunk* c, const char* symbol) {
     int n_packs = (len + 7) / 8;
     chunk_emit(c, OP_CONST, chunk_add_const(c, INT_VAL(len)));
     for (int p = 0; p < n_packs; p++) {
-        int64_t pack = 0;
+        uint64_t pack = 0;
         for (int b = 0; b < 8 && p * 8 + b < len; b++)
-            pack |= ((int64_t)(unsigned char)symbol[p * 8 + b]) << (b * 8);
-        chunk_emit(c, OP_CONST, chunk_add_const(c, INT_VAL(pack)));
+            pack |= ((uint64_t)(unsigned char)symbol[p * 8 + b]) << (b * 8);
+        chunk_emit(c, OP_CONST, chunk_add_const(c, INT_VAL((int64_t)pack)));
     }
     chunk_emit(c, OP_NATIVE_CALL,
                ESHKOL_VM_PACKED_SYMBOL_FID_BASE + n_packs);
@@ -1504,11 +1504,11 @@ static void compile_form_define_record_type(FuncChunk* c, Node* node, int tail) 
         int n_packs = (len + 7) / 8;
         chunk_emit(&func, OP_CONST, chunk_add_const(&func, INT_VAL(len)));
         for (int p = 0; p < n_packs; p++) {
-            int64_t pack = 0;
+            uint64_t pack = 0;
             for (int b = 0; b < 8 && p * 8 + b < len; b++) {
-                pack |= ((int64_t)(unsigned char)node->children[1]->symbol[p * 8 + b]) << (b * 8);
+                pack |= ((uint64_t)(unsigned char)node->children[1]->symbol[p * 8 + b]) << (b * 8);
             }
-            chunk_emit(&func, OP_CONST, chunk_add_const(&func, INT_VAL(pack)));
+            chunk_emit(&func, OP_CONST, chunk_add_const(&func, INT_VAL((int64_t)pack)));
         }
         chunk_emit(&func, OP_NATIVE_CALL,
                    ESHKOL_VM_PACKED_STRING_FID_BASE + n_packs);
@@ -3464,11 +3464,11 @@ static void compile_expr_impl(FuncChunk* c, Node* node, int tail) {
         int n_packs = (len + 7) / 8;
         chunk_emit(c, OP_CONST, chunk_add_const(c, INT_VAL(len)));
         for (int p = 0; p < n_packs; p++) {
-            int64_t pack = 0;
+            uint64_t pack = 0;
             for (int b = 0; b < 8 && p * 8 + b < len; b++) {
-                pack |= ((int64_t)(unsigned char)node->string_data[p * 8 + b]) << (b * 8);
+                pack |= ((uint64_t)(unsigned char)node->string_data[p * 8 + b]) << (b * 8);
             }
-            chunk_emit(c, OP_CONST, chunk_add_const(c, INT_VAL(pack)));
+            chunk_emit(c, OP_CONST, chunk_add_const(c, INT_VAL((int64_t)pack)));
         }
         chunk_emit(c, OP_NATIVE_CALL,
                    ESHKOL_VM_PACKED_STRING_FID_BASE + n_packs);
